Implemented new controller - player in game architecture and networked movement
Had to refactor the default pawn to be a floating invisible pawn. This
pawn is possessed by the player controller. The player controller then
spawns a Top Down Character which is possessed by a PAIController. A
reference to this PAIController is maintained in the player controller,
which is used to move the player