Initial implementation of death.
There was some reworking of respawing:
Interactable::DummyRespawn() -> Unit::RespawnCaller()
Recall this client-side function goes on to call the server-side PlayerController event.

Moved variables involving death from MobaCharacter to Interactable as it makes more sense that anything with the ability to 'die' should keep track of said state.
Following that up, Interactable::Die() is a new function which is called any time RemainingHealth <= 0.